Remove :start action from neo4j service to get it to startup on initial convergance on el6
When converging for the first time on EL6 systems, the neo4j service is not
running by default. This appears to be caused by the same issues noted in
http://tickets.opscode.com/browse/COOK-1599. I have applied the same fix
as they chose, removing the :start action from the neo4j service and
depending on the notification from config file template updates to
start the service during the initial convergence.